third_party/blink/web_tests/fast/events/page-scaled-mouse-click-expected.txt

This tests that page scaling does not affect mouse event pageX and pageY coordinates.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Unscaled
PASS eventLog is "click(10, 10)"


setPageScale(0.5)
PASS eventLog is "click(20, 20)"


PASS successfullyParsed is true

TEST COMPLETE
